bdcmc now matches bdcmc-test

+++ src/util/hardware_adc.hpp
@@ -22,12 +22,13 @@ namespace mrover {
 			m_values.resize(channels);
 		}
 
-		uint32_t get_raw_channel_value(uint8_t channel) {
+		uint16_t get_raw_channel_value(uint8_t channel) {
 			return m_values.at(channel);
 		}
 
 		void update() {
 
+//			HAL_ADC_Start_DMA(m_hadc);
 			HAL_ADC_Start_DMA(m_hadc, reinterpret_cast<uint32_t*>(m_values.data()), m_channels / (sizeof(uint32_t) / sizeof(uint16_t)));
 		}
